+++ previous-versions/common/pkg/models/operations/updateconnectionauthorization.go
@@ -41,6 +41,8 @@ type UpdateConnectionAuthorizationResponse struct {
 	Connection *shared.Connection
 	// HTTP response content type for this operation
 	ContentType string
+	// Your API request was not properly authorized.
+	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
@@ -59,6 +61,13 @@ func (o *UpdateConnectionAuthorizationResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
+}
+
+func (o *UpdateConnectionAuthorizationResponse) GetErrorMessage() *shared.ErrorMessage {
+	if o == nil {
+		return nil
+	}
+	return o.ErrorMessage
 }
 
 func (o *UpdateConnectionAuthorizationResponse) GetStatusCode() int {